timeutils: move get_local_timezone_ofs() from misc to cache.c
This function is actually a cache and we will need this in a subsequent
patch from within cache.c.
