+++ lib/timeutils/cache.c
@@ -102,6 +102,44 @@ static struct
 
 static GMutex localtime_lock;
 
+/**
+ * get_local_timezone_ofs:
+ * @when: time in UTC
+ *
+ * Return the zone offset (measured in seconds) of @when expressed in local
+ * time. The function also takes care about daylight saving.
+ **/
+long
+get_local_timezone_ofs(time_t when)
+{
+#ifdef SYSLOG_NG_HAVE_STRUCT_TM_TM_GMTOFF
+  struct tm ltm;
+
+  cached_localtime(&when, &ltm);
+  return ltm.tm_gmtoff;
+
+#else
+
+  struct tm gtm;
+  struct tm ltm;
+  long tzoff;
+
+  cached_localtime(&when, &ltm);
+  cached_gmtime(&when, &gtm);
+
+  tzoff = (ltm.tm_hour - gtm.tm_hour) * 3600;
+  tzoff += (ltm.tm_min - gtm.tm_min) * 60;
+  tzoff += ltm.tm_sec - gtm.tm_sec;
+
+  if (tzoff > 0 && (ltm.tm_year < gtm.tm_year || ltm.tm_mon < gtm.tm_mon || ltm.tm_mday < gtm.tm_mday))
+    tzoff -= 86400;
+  else if (tzoff < 0 && (ltm.tm_year > gtm.tm_year || ltm.tm_mon > gtm.tm_mon || ltm.tm_mday > gtm.tm_mday))
+    tzoff += 86400;
+
+  return tzoff;
+#endif /* SYSLOG_NG_HAVE_STRUCT_TM_TM_GMTOFF */
+}
+
 static glong
 _get_system_tzofs(void)
 {
