+++ modules/timestamp/tests/test_format_date.c
@@ -23,7 +23,7 @@
 #include "libtest/cr_template.h"
 
 #include "logmsg/logmsg.h"
-#include "timeutils/misc.h"
+#include "timeutils/cache.h"
 #include "apphook.h"
 #include "scratch-buffers.h"
 #include "cfg.h"
